document-analysis-app/requirements.txt
Flask
pandas
numpy
scikit-learn
transformers
beautifulsoup4
requests
tqdm
# Document processing dependencies
PyPDF2
pdfplumber
markdown
python-docx
# Multi-LLM API dependencies
openai
google-generativeai
anthropic
ollama
# LangChain dependencies
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-ollama
unstructured
# MCP Server dependencies (optional - for MCP server functionality)
# mcp>=1.0.0  # Uncomment if MCP server functionality is needed
pydantic>=2.0.0
# Additional utilities
pathlib2
python-dotenv
argparse
uuid